feat(database): implement Prisma ORM with PostgreSQL integration
- Integrated Prisma ORM with PostgreSQL for robust database management
- Configured TypeScript with proper type checking and module resolution
  - Set up tsconfig with strict mode and path aliases
  - Added @types/node for Node.js type definitions
- Added development tooling
  - Integrated nodemon for hot-reloading during development
  - Configured ts-node for TypeScript execution
- Implemented database connection
  - Created shared Prisma client instance
  - Set up environment variable configuration
  - Added database connection test script
- Documented setup process
  - Created comprehensive Prisma setup guide
  - Added troubleshooting section for common issues
  - Documented database management commands
- Separated concerns
  - Moved Prisma configuration to dedicated files
  - Isolated database logic from business logic
- Added TypeScript path aliases for cleaner imports
- Set up database migrations with proper schema versioning

The database layer is now ready for development with proper type safety and development tooling in place. The setup includes comprehensive documentation for future developers.
