+++ Web-Api/.env.example
@@ -1,0 +1,2 @@
+DEBUG=true/false
+JWT_SECRET=your_jwt_secret
