+++ Web-Api/nodemon.json
@@ -1,0 +1,6 @@
+{
+    "watch": ["src"],
+    "ext": "ts",
+    "ignore": ["dist"],
+    "exec": "ts-node -r tsconfig-paths/register src/server.ts"
+  }
