+++ Web-Api/src/errors/Unauthorized.ts
@@ -1,6 +1,6 @@
 export class Unauthorized extends Error {
-  constructor(message = 'Unauthorized') {
-    super(message);
-    Object.setPrototypeOf(this, Unauthorized.prototype);
-  }
+    constructor(message = 'Unauthorized') {
+        super(message);
+        Object.setPrototypeOf(this, Unauthorized.prototype);
+    }
 }
