+++ Web-Api/src/errors/validation.ts
@@ -1,8 +1,8 @@
 export class Validation extends Error {
-  public e: any;
-  constructor(e: any, message = 'Validation failed') {
-    super(message);
-    this.e = e;
-    Object.setPrototypeOf(this, Validation.prototype);
-  }
+    public e: any;
+    constructor(e: any, message = 'Validation failed') {
+        super(message);
+        this.e = e;
+        Object.setPrototypeOf(this, Validation.prototype);
+    }
 }
