+++ Web-Api/src/test-db.ts
@@ -1,0 +1,31 @@
+import dotenv from 'dotenv';
+import path from 'path';
+
+// Load environment variables from .env file
+dotenv.config({ path: path.resolve(__dirname, '../.env') });
+
+import prisma from './prisma';
+
+async function testConnection() {
+  try {
+    console.log('Using database URL:', process.env.DATABASE_URL?.split('@')[1] || 'Not found');
+    
+    // Test connection by querying the database
+    const result = await prisma.$queryRaw`SELECT 1 as test`;
+    console.log('✅ Database connection successful!', result);
+    
+    // Try to list tables (if any)
+    const tables = await prisma.$queryRaw`
+      SELECT table_name 
+      FROM information_schema.tables 
+      WHERE table_schema = 'public'
+    `;
+    console.log('📊 Tables in database:', tables);
+  } catch (error) {
+    console.error('❌ Database connection failed:', error);
+  } finally {
+    await prisma.$disconnect();
+  }
+}
+
+testConnection();
