+++ Web-Api/src/utils/authUtils.ts
@@ -7,21 +7,21 @@ const JWT_SECRET = process.env.JWT_SECRET || 'default_jwt_secret';
 const JWT_EXPIRES_IN = '1h';
 
 export const hashPassword = async (password: string): Promise<string> => {
-  return bcrypt.hash(password, SALT_ROUNDS);
+    return bcrypt.hash(password, SALT_ROUNDS);
 };
 
 export const comparePassword = async (password: string, hash: string): Promise<boolean> => {
-  return bcrypt.compare(password, hash);
+    return bcrypt.compare(password, hash);
 };
 
 export const generateApiKey = (): string => {
-  return 'sk_' + crypto.randomBytes(32).toString('hex');
+    return 'sk_' + crypto.randomBytes(32).toString('hex');
 };
 
 export const generateJWT = (payload: object): string => {
-  return jwt.sign(payload, JWT_SECRET, { expiresIn: JWT_EXPIRES_IN });
+    return jwt.sign(payload, JWT_SECRET, { expiresIn: JWT_EXPIRES_IN });
 };
 
 export const verifyJWT = (token: string): any => {
-  return jwt.verify(token, JWT_SECRET);
+    return jwt.verify(token, JWT_SECRET);
 };
